+++ protocol/state/blockindex.go
@@ -44,10 +44,10 @@ func NewBlockNode(bh *types.BlockHeader, parent *BlockNode) (*BlockNode, error) 
 		TransactionStatusHash:  bh.TransactionStatusHash,
 	}
 
-	node.BlockWitness = common.NewBitMap(uint64(len(bh.Witness)))
+	node.BlockWitness = common.NewBitMap(uint32(len(bh.Witness)))
 	for i, witness := range bh.Witness {
 		if len(witness) != 0 {
-			node.BlockWitness.Set(uint64(i))
+			node.BlockWitness.Set(uint32(i))
 		}
 	}
 	return node, nil
